# openapi.yaml
openapi: 3.0.0
info:
  title: Video Transcription Service API
  description: Microservice for video transcription using Azure AI Speech Services
  version: 1.0.0
  contact:
    name: eUčilnica+ Team
    url: https://github.com/Trije-bingusi

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Health
    description: Health check and readiness endpoints
  - name: Metrics
    description: Prometheus metrics
  - name: Transcriptions
    description: Video transcription management

paths:
  /healthz:
    get:
      summary: Health check
      description: Returns OK if the service is running
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            text/plain:
              schema:
                type: string
                example: OK

  /readyz:
    get:
      summary: Readiness check
      description: Returns READY if the service can accept requests (DB is connected)
      tags:
        - Health
      responses:
        '200':
          description: Service is ready
          content:
            text/plain:
              schema:
                type: string
                example: READY
        '500':
          description: Service is not ready
          content:
            text/plain:
              schema:
                type: string
                example: NOT READY

  /metrics:
    get:
      summary: Prometheus metrics
      description: Returns Prometheus metrics for monitoring
      tags:
        - Metrics
      responses:
        '200':
          description: Metrics in Prometheus format
          content:
            text/plain:
              schema:
                type: string

  /api/lectures/{lectureId}/transcriptions:
    get:
      summary: Get all transcriptions for a lecture
      description: Returns a list of all transcription jobs associated with a specific lecture
      tags:
        - Transcriptions
      parameters:
        - name: lectureId
          in: path
          required: true
          description: UUID of the lecture
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of transcriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranscriptionSummary'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/lectures/{lectureId}/transcribe:
    post:
      summary: Create a new transcription job
      description: Starts a new transcription job for a video
      tags:
        - Transcriptions
      parameters:
        - name: lectureId
          in: path
          required: true
          description: UUID of the lecture
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                video_url:
                  type: string
                  description: URL to the video file
                  example: https://example.blob.core.windows.net/videos/lecture1.mp4
                language:
                  type: string
                  description: Language code for transcription (BCP-47 format)
                  default: en-US
                  example: en-US
              required:
                - video_url
      responses:
        '201':
          description: Transcription job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transcription'
        '400':
          description: Bad request (missing video_url)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/transcriptions/{id}:
    get:
      summary: Get a specific transcription
      description: Returns full details of a transcription including word timings
      tags:
        - Transcriptions
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the transcription
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transcription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transcription'
        '404':
          description: Transcription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a transcription
      description: Deletes a transcription job and all associated data
      tags:
        - Transcriptions
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the transcription
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transcription deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Transcription deleted successfully
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/transcriptions/{id}/transcript:
    get:
      summary: Get transcript text
      description: Returns the plain text transcript
      tags:
        - Transcriptions
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the transcription
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Transcript text
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcript:
                    type: string
                    description: Full transcript text
        '400':
          description: Transcription not yet completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Transcription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/transcriptions/{id}/vtt:
    get:
      summary: Get VTT subtitles
      description: Returns WebVTT format subtitles with timestamps
      tags:
        - Transcriptions
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the transcription
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: VTT subtitle file
          content:
            text/vtt:
              schema:
                type: string
                example: |
                  WEBVTT

                  00:00:00.000 --> 00:00:02.500
                  This is the first subtitle

                  00:00:02.500 --> 00:00:05.000
                  This is the second subtitle
        '400':
          description: Transcription not yet completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Transcription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/transcriptions/{id}/srt:
    get:
      summary: Get SRT subtitles
      description: Returns SubRip (SRT) format subtitles with timestamps
      tags:
        - Transcriptions
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the transcription
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: SRT subtitle file
          content:
            text/plain:
              schema:
                type: string
                example: |
                  1
                  00:00:00,000 --> 00:00:02,500
                  This is the first subtitle

                  2
                  00:00:02,500 --> 00:00:05,000
                  This is the second subtitle
        '400':
          description: Transcription not yet completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Transcription not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Transcription:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the transcription
        lecture_id:
          type: string
          format: uuid
          description: ID of the associated lecture
        video_url:
          type: string
          description: URL to the video file
        language:
          type: string
          description: Language code (BCP-47 format)
          example: en-US
        status:
          type: string
          enum: [pending, processing, completed, failed]
          description: Status of the transcription job
        azure_job_id:
          type: string
          nullable: true
          description: Azure Speech Services job ID
        transcript:
          type: string
          nullable: true
          description: Full transcript text
        word_timings:
          type: array
          nullable: true
          description: Word-level timing information
          items:
            type: object
            properties:
              word:
                type: string
              start:
                type: number
                description: Start time in seconds
              end:
                type: number
                description: End time in seconds
              confidence:
                type: number
                description: Confidence score (0-1)
        vtt_content:
          type: string
          nullable: true
          description: WebVTT format subtitles
        srt_content:
          type: string
          nullable: true
          description: SubRip format subtitles
        error_message:
          type: string
          nullable: true
          description: Error message if transcription failed
        duration:
          type: integer
          nullable: true
          description: Video duration in seconds
        confidence:
          type: number
          nullable: true
          description: Average confidence score
        created_at:
          type: string
          format: date-time
          description: Timestamp when job was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when job was last updated
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when job completed

    TranscriptionSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lecture_id:
          type: string
          format: uuid
        video_url:
          type: string
        language:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        confidence:
          type: number
          nullable: true
        duration:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message